Enable converion warnings for GCC and clang
Past me didn't want to deal with fixing warnings related to comparing
the size of a container and the default int. Now that we are using
appropriate types, this is simpler to fix.

There are some static_casts which may be quick fixes so might need
revisiting in the future.

+++ include/nuclear-data-reader/converter.hpp
@@ -97,7 +97,7 @@ public:
     uint64_t c{};
     for (auto p = s; *p != 0; ++p, ++c)
       {
-        hash += *p << c;
+        hash += static_cast<uint64_t>(*p) << c;
       }
     return hash;
   }
